Time: 0 ms (100%), Space: 8.5 MB (83.59%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -16,6 +16,6 @@ public:
         if(rev > INT_MAX || rev < INT_MIN)
             return 0;
 
-        return (int)rev;
+        return rev;
     }
 };
